+++ pkg/vm/engine/tae/dataio/segmentio/block.go
@@ -29,7 +29,7 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/container/compute"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/container/vector"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/iface/file"
-	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/tables/indeximpl"
+	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/tables/indexwrapper"
 )
 
 type blockFile struct {
@@ -119,7 +119,7 @@ func (bf *blockFile) LoadIndexMeta() (any, error) {
 	if err != nil {
 		return nil, err
 	}
-	indices := indeximpl.NewEmptyIndicesMeta()
+	indices := indexwrapper.NewEmptyIndicesMeta()
 	if err = indices.Unmarshal(buf); err != nil {
 		return nil, err
 	}
