+++ pkg/vm/engine/tae/tables/indexwrapper/immutindex.go
@@ -1,4 +1,4 @@
-package indeximpl
+package indexwrapper
 
 import (
 	"github.com/RoaringBitmap/roaring"
