+++ pkg/vm/engine/tae/tables/indexwrapper/meta.go
@@ -1,4 +1,4 @@
-package indeximpl
+package indexwrapper
 
 import (
 	"bytes"
